app/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
PyJWT==2.9.0

